feat: translations
fixing the labels and subtype editor
