+++ src/pages/TheCrucible/Profession/components/SubtypeEditor.tsx
@@ -2,8 +2,10 @@ import { useState } from 'preact/hooks';
 import { Button, TextInput } from 'react95';
 import styled from 'styled-components';
 import { t } from '@lingui/core/macro';
+import { i18n } from '@lingui/core';
 
 import { IoPencilOutline, IoCheckmarkSharp } from "react-icons/io5";
+import { DEFAULT_SKILLS } from '../../../../types/characterTypes';
 import { useSkills } from '../../../../providers/SkillsContext';
 import Dialogue from '../../../../components/Dialogue/Dialogue';
 import { Skill } from '../../../../types/characterTypes';
@@ -60,8 +62,10 @@ interface SubtypeEditorProps {
 function SubtypeEditor({ skill }: SubtypeEditorProps) {
     const { setSkillById } = useSkills();
     const [showModal, setShowModal] = useState(false);
-    // AJS start here, show translated subtype if we're on default subtype
-    const [localSubType, setLocalSubType] = useState(skill.subType || '');
+    const [localSubType, setLocalSubType] = useState(() => {
+        const defaultSkill = DEFAULT_SKILLS.find(s => s.id === skill.id);
+        return defaultSkill?.subType === skill.subType ? i18n._(skill.subTypeMsg) : skill.subType || '';
+    });
 
     const handleSubtypeChange = (e: any) => {
         setLocalSubType(e?.target?.value);
